+++ apps/site/app/page.tsx
@@ -1,5 +1,0 @@
-import { OfficialSite } from "../src/OfficialSite";
-
-export default function Home() {
-  return <OfficialSite />;
-}